requirements.txt
marko
pillow